apps/api/pyproject.toml
[tool.poetry]
name = "grantd-api"
version = "0.1.0"
description = "Grantd API - Visual RBAC for data platforms"
authors = ["Mikkel Gronning <mikkel@grantd.io>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
mangum = "^0.18.0"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.36"
boto3 = "^1.35.0"
snowflake-connector-python = "^3.12.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
httpx = "^0.28.0"
pydantic = "^2.10.0"
pydantic-settings = "^2.6.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
ruff = "^0.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
